add functional module
