+++ graphgallery/data_type.py
@@ -15,7 +15,7 @@ __all__ = ['is_iterable',
            'infer_type',
            'is_tensor',
            'is_strided_tensor',
-           'is_sparse_tensor',
+           'is_sparse',
            ]
 
 
@@ -26,11 +26,9 @@ def is_iterable(obj: Any) -> bool:
 
 def is_listlike(x: Any) -> bool:
     """Check whether `x` is list like, e.g., Tuple or List.
-
     Parameters:
     ----------
     x: A python object to check.
-
     Returns:
     ----------
     `True` iff `x` is a list like sequence.
@@ -38,27 +36,23 @@ def is_listlike(x: Any) -> bool:
     return isinstance(x, (list, tuple))
 
 
-
 def is_scalar(x: Any) -> bool:
     """Check whether `x` is a scalar, an array scalar, or a 0-dim array.
-
     Parameters:
     ----------
     x: A python object to check.
-
     Returns:
     ----------
     `True` iff `x` is a scalar, an array scalar, or a 0-dim array.
     """
     return np.isscalar(x) or (isinstance(x, np.ndarray) and x.ndim == 0)
 
+
 def is_intscalar(x: Any) -> bool:
     """Check whether `x` is an integer scalar.
-
     Parameters:
     ----------
     x: A python object to check.
-
     Returns:
     ----------
     `True` iff `x` is an integer scalar (built-in or Numpy integer).
@@ -76,11 +70,9 @@ def is_intscalar(x: Any) -> bool:
 
 def is_floatscalar(x: Any) -> bool:
     """Check whether `x` is a float scalar.
-
     Parameters:
     ----------
     x: A python object to check.
-
     Returns:
     ----------
     `True` iff `x` is a float scalar (built-in or Numpy float).
@@ -90,22 +82,19 @@ def is_floatscalar(x: Any) -> bool:
                           np.float32,
                           np.float64,
                           ))
-    
+
 
 def infer_type(x: Any) -> str:
     """Infer type of the input `x`.
-
     Parameters:
     ----------
     x: Any python object
-
     Returns:
     ----------
     dtype: string, the proper data type of `x`:
         1. `graphgallery.floatx()` if `x` is floating,
         2. `graphgallery.intx()` if `x` is integer,
         3. `graphgallery.boolx()` if `x` is boolean.
-
     """
     # For tensor or variable
     if is_th_tensor(x):
@@ -117,7 +106,7 @@ def infer_type(x: Any) -> str:
             return gg.intx()
         else:
             raise RuntimeError(f'Invalid input of `{type(x)}`')
-        
+
     elif is_tf_tensor(x):
         if x.dtype.is_floating:
             return gg.floatx()
@@ -141,19 +130,18 @@ def infer_type(x: Any) -> str:
         raise RuntimeError(f'Invalid inputs of `{x}`.')
     else:
         raise RuntimeError(f'Invalid input of `{type(x)}`')
-    
 
-def is_sparse_tensor(x: Any, backend: Optional[str] = None) -> bool:
+
+def is_sparse(x: Any, backend: Optional[str] = None) -> bool:
     """Check whether `x` is a sparse Tensor.
-    
+
     Parameters:
     ----------
     x: A python object to check.
-    
+
     backend: String or 'BackendModule', optional.    
      `'tensorflow'`, `'torch'`, TensorFlowBackend, PyTorchBackend, etc.    
      if not specified, return the current default backend module. 
-
     Returns:
     ----------
     `True` iff `x` is a (tf or torch) sparse-tensor.
@@ -168,29 +156,29 @@ def is_sparse_tensor(x: Any, backend: Optional[str] = None) -> bool:
 
 def is_strided_tensor(x: Any, backend: Optional[str] = None) -> bool:
     """Check whether `x` is a strided (dense) Tensor.
-    
+
     Parameters:
     ----------
     x: A python object to check.
-    
+
     backend: String or 'BackendModule', optional.    
      `'tensorflow'`, `'torch'`, TensorFlowBackend, PyTorchBackend, etc.    
      if not specified, return the current default backend module. 
-     
+
     Returns:
     ----------
     `True` iff `x` is a (tf or torch) strided (dense) Tensor.
     """
-    
+
     backend = gg.backend(backend)
 
     if backend == 'tensorflow':
         return is_tf_strided_tensor(x)
     else:
         return is_th_strided_tensor(x)
-    
 
-def is_tensor(x: Any, backend: Optional[str]=None) -> bool:
+
+def is_tensor(x: Any, backend: Optional[str] = None) -> bool:
     """Check whether `x` is 
         tf.Tensor,
         tf.Variable,
@@ -198,15 +186,13 @@ def is_tensor(x: Any, backend: Optional[str]=None) -> bool:
         tf.sparse.SparseTensor,
         torch.Tensor, 
         torch.sparse.Tensor.
-
     Parameters:
     ----------
     x: A python object to check.
-    
+
     backend: String or 'BackendModule', optional.    
      `'tensorflow'`, `'torch'`, TensorFlowBackend, PyTorchBackend, etc.    
      if not specified, return the current default backend module.    
-
     Returns:
     ----------
     `True` iff `x` is a (tf or torch) (sparse-)tensor.
@@ -235,9 +221,11 @@ def is_tf_strided_tensor(x: Any) -> bool:
 
 def is_th_strided_tensor(x: Any) -> bool:
     return is_th_tensor(x) and x.layout == torch.strided
-               
+
+
 def is_tf_tensor(x: Any) -> bool:
     return is_tf_strided_tensor(x) or is_tf_sparse_tensor(x)
+
 
 def is_th_tensor(x: Any) -> bool:
     return torch.is_tensor(x)
