+++ graphgallery/functional/__init__.py
@@ -1,0 +1,3 @@
+from .ops import *
+from .tensor import *
+from .device import parse_device
