+++ graphgallery/functional/torch/__init__.py
@@ -1,0 +1,3 @@
+from .device import *
+from .tensor import *
+from .sparse import *
