+++ graphgallery/nn/models/semisupervised/gat.py
@@ -13,6 +13,7 @@ from graphgallery.utils.decorators import EqualVarLength
 
 from graphgallery.nn.models.semisupervised.th_models.gat import GAT as pyGAT
 from graphgallery.nn.models.semisupervised.tf_models.gat import GAT as tfGAT
+from graphgallery import functional as F
 from graphgallery import transforms as T
 
 
@@ -75,34 +76,33 @@ class GAT(SemiSupervisedModel):
         adj_matrix = self.adj_transform(graph.adj_matrix)
         attr_matrix = self.attr_transform(graph.attr_matrix)
 
-        self.feature_inputs, self.structure_inputs = T.astensors(
+        self.feature_inputs, self.structure_inputs = F.astensors(
             attr_matrix, adj_matrix, device=self.device)
 
     @EqualVarLength(include=["n_heads"])
-    def build(self, hiddens=[8], n_heads=[8], activations=['elu'], 
+    def build(self, hiddens=[8], n_heads=[8], activations=['elu'],
               dropout=0.6, l2_norm=5e-4,
               lr=0.01, use_bias=True):
 
         if self.backend == "tensorflow":
             with tf.device(self.device):
-                self.model = tfGAT(self.graph.n_attrs, self.graph.n_classes, 
+                self.model = tfGAT(self.graph.n_attrs, self.graph.n_classes,
                                    hiddens=hiddens, n_heads=n_heads,
-                                   activations=activations, 
-                                   dropout=dropout, 
+                                   activations=activations,
+                                   dropout=dropout,
                                    l2_norm=l2_norm,
                                    lr=lr, use_bias=use_bias)
 
         else:
-            self.model = pyGAT(self.graph.n_attrs, self.graph.n_classes, 
+            self.model = pyGAT(self.graph.n_attrs, self.graph.n_classes,
                                hiddens=hiddens, n_heads=n_heads,
-                               activations=activations, 
-                               dropout=dropout, 
+                               activations=activations,
+                               dropout=dropout,
                                l2_norm=l2_norm,
                                lr=lr, use_bias=use_bias).to(self.device)
 
-
     def train_sequence(self, index):
-        
+
         labels = self.graph.labels[index]
         sequence = FullBatchNodeSequence(
             [self.feature_inputs, self.structure_inputs, index], labels, device=self.device)
