+++ graphgallery/nn/models/semisupervised/gwnn.py
@@ -7,7 +7,7 @@ from graphgallery.utils.decorators import EqualVarLength
 from graphgallery.nn.models.semisupervised.tf_models.gwnn import GWNN as tfGWNN
 
 from graphgallery import transforms as T
-
+from graphgallery import functional as F
 
 
 class GWNN(SemiSupervisedModel):
@@ -70,7 +70,7 @@ class GWNN(SemiSupervisedModel):
         adj_matrix = self.adj_transform(graph.adj_matrix)
         attr_matrix = self.attr_transform(graph.attr_matrix)
 
-        self.feature_inputs, self.structure_inputs = T.astensors(
+        self.feature_inputs, self.structure_inputs = F.astensors(
             attr_matrix, adj_matrix, device=self.device)
 
     # use decorator to make sure all list arguments have the same length
@@ -89,7 +89,7 @@ class GWNN(SemiSupervisedModel):
             raise NotImplementedError
 
     def train_sequence(self, index):
-        
+
         labels = self.graph.labels[index]
         sequence = FullBatchNodeSequence(
             [self.feature_inputs, *self.structure_inputs, index], labels, device=self.device)
