+++ graphgallery/nn/models/semisupervised/robustgcn.py
@@ -6,6 +6,7 @@ from graphgallery.utils.decorators import EqualVarLength
 
 from graphgallery.nn.models.semisupervised.tf_models.robustgcn import RobustGCN as tfRobustGCN
 from graphgallery import transforms as T
+from graphgallery import functional as F
 
 
 class RobustGCN(SemiSupervisedModel):
@@ -65,7 +66,7 @@ class RobustGCN(SemiSupervisedModel):
         adj_matrix = self.adj_transform(graph.adj_matrix)
         attr_matrix = self.attr_transform(graph.attr_matrix)
 
-        self.feature_inputs, self.structure_inputs = T.astensors(
+        self.feature_inputs, self.structure_inputs = F.astensors(
             attr_matrix, adj_matrix, device=self.device)
 
     # use decorator to make sure all list arguments have the same length
@@ -85,7 +86,7 @@ class RobustGCN(SemiSupervisedModel):
             raise NotImplementedError
 
     def train_sequence(self, index):
-        
+
         labels = self.graph.labels[index]
         sequence = FullBatchNodeSequence(
             [self.feature_inputs, *self.structure_inputs, index], labels, device=self.device)
