+++ graphgallery/transforms/__init__.py
@@ -9,8 +9,8 @@ from graphgallery.transforms.edge_transform.edge_transpose import edge_transpose
 from graphgallery.transforms.edge_transform.add_selfloops_edge import add_selfloops_edge
 from graphgallery.transforms.edge_transform.augment_edge import augment_edge
 from graphgallery.transforms.edge_transform.normalize_edge import normalize_edge
-from graphgallery.transforms.edge_transform.edges2adj import (sparse_edges_to_sparse_adj,
-                                                                  SparseEdgesToSparseAdj)
+from graphgallery.transforms.edge_transform.edges2adj import (sparse_edge_to_sparse_adj,
+                                                              SparseEdgesToSparseAdj)
 # adj_transform
 from graphgallery.transforms.graph_partition import GraphPartition, graph_partition
 from graphgallery.transforms.adj_transform.normalize_adj import NormalizeAdj, normalize_adj
@@ -20,7 +20,7 @@ from graphgallery.transforms.adj_transform.chebyshef import ChebyBasis, cheby_ba
 from graphgallery.transforms.adj_transform.neighbor_sampler import NeighborSampler, neighbor_sampler
 from graphgallery.transforms.adj_transform.gdc import GDC, gdc
 from graphgallery.transforms.adj_transform.svd import SVD, svd
-from graphgallery.transforms.adj_transform.adj2edges import sparse_adj_to_sparse_edges, SparseAdjToSparseEdges
+from graphgallery.transforms.adj_transform.adj2edges import sparse_adj_to_sparse_edge, SparseAdjToSparseEdges
 from graphgallery.transforms.adj_transform.augment_adj import augment_adj
 from graphgallery.transforms.adj_transform.sparse_reshape import SparseReshape, sparse_reshape
 
@@ -28,12 +28,6 @@ from graphgallery.transforms.adj_transform.sparse_reshape import SparseReshape, 
 # attr_transform
 from graphgallery.transforms.attr_transform.normalize_attr import NormalizeAttr, normalize_attr
 from graphgallery.transforms.attr_transform.augment_attr import augment_attr
-
-
-
-# tensor transform
-from graphgallery.transforms.tensor_transform.tensor import *
-from graphgallery.transforms.tensor_transform import tf_tensor, th_tensor
 
 
 # other transform
