--- graphgallery/transforms/tensor_transform/__init__.py
@@ -1,1 +1,0 @@
-from graphgallery.transforms.tensor_transform.tensor import *
